Modified the filter API logi & simplified the mycourse_controller.js code by abstracting functions into another file
